Move function with only one usage
