+++ client/go/internal/admin/jvm/application_container_test.go
@@ -1,5 +1,7 @@
 // Copyright Vespa.ai. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package util
+// Author: arnej
+
+package jvm
 
 import (
 	"testing"
@@ -8,6 +10,6 @@ import (
 )
 
 func TestMD5SimpleInputs(t *testing.T) {
-	assert.Equal(t, "d41d8cd98f00b204e9800998ecf8427e", Md5Hex(""))
-	assert.Equal(t, "4044e8209f286312a68bbb54f8714922", Md5Hex("admin/cluster-controllers/0\n"))
+	assert.Equal(t, "d41d8cd98f00b204e9800998ecf8427e", md5Hex(""))
+	assert.Equal(t, "4044e8209f286312a68bbb54f8714922", md5Hex("admin/cluster-controllers/0\n"))
 }
